added trigger management to tessim, various other small C fixes
tessim now has the capability to trigger on events. This is done using
the new triggertype parameter, which defines the trigger algorithm to use:

triggertype=stream  -- write a TESDATASTREAM file
triggertype=movavg:npts:threshold:suppress
    trigger based on a moving average of npts points. If the current
    (digitized) pulse is larger than threshold*moving average, then we
    trigger. The optional parameter suppress tells the trigger logic
    not to trigger again for suppres samples. So, for example:
    triggertype=movavg:5:15.:10 will trigger if the current point is
    15 times larger than the average of the previous 5 samples.

triggertype=diff:npts:threshold:suppress
    trigger based on the low pass filtered derivative based on the
    previous npts points. npts can be 2,3,4,5,6,7,8,9 or 10. The left
    handed derivative equations from
    http://www.holoborodko.com/pavel/numerical-methods/numerical-derivative/smooth-low-noise-differentiators/
    are used. These are of high quality, I find that in practice
    npts=4 is sufficient. The other parameters are the same as for the
    movavg trigger.

If a trigger is used, tessim will write a FITS-file containing the
individual records. The total length of the trigger is set by the
new triggersize parameter, while the new prebuffer parameter contains
the number of samples before the trigger. So, if triggersize=1024 and
prebuffer=100 then the record written to file will have a length of
1024 and will start 100 samples before the trigger happened.

Note that if a second trigger happens while we're already accumulating
data, the size of the record will be increased in order to accomodate
this secondary.

This commit also contains various smaller fixes in other routines
(mainly small C optimizations/code simplifications)