Implemented possibility to define different RMF matrices for different gradings in advdet XML description.
WARNING : after this commit, the pixrmf tag will be considered invalid.
The way to define several gradings for a pixel and its corresponding RMFs is now as follows:
<grading num="1" name="high" pre="128" post="512" rmf="athena_xifu_rmf_highres_v20150609.rmf"/>
<grading num="2" name="mid" pre="128" post="128" rmf="athena_xifu_rmf_midres_v20150609.rmf"/>
<grading num="3" name="low" pre="128" post="0" rmf="athena_xifu_rmf_lowres_v20150609.rmf"/>

These gradings can be defined either inside the <pixel> and <pixel/> tags, in which case they affect only this pixel, or outside in which case they define the grading for all previously defined pixels.