[major update to version 2.1.0]
Previously the output products of simulations were not named
consistently and resulted in confusion for the user. From v2.1.0 on
those will be standardized in the following way:

     ** EvtFile ** is now the final (cleaned) product of the simulations,
     often also referred to as level-2-event-file. Previously this was
     known as PatternList in SIXTE

     ** RawData ** are now the intermediate and raw data products of a
     simulation. Previously this was often called EventList in SIXTE.

This change will break backwards compatibility and will require to
rename these keywords in your scripts. Currently you will get a
notification if you use the old keywords and you will be pointed to the
new name.

Additinally the tool for simulating the complete four athena chips has
been renamed to 'athenawfisim'.