Implementing the NXB in xifupipeline The new NXB file will be provided shortly It will correspond to the requirement of 5e-3 cts/s/keV/cm2 The corresponding pha has this value in cts/s/bin/m2 as required To activate NXB use key argument "Background=yes"
--> Photon generation part of xifupipeline changed to account for NXB.
Creation of a new routine to generate all the photons of the NXB (very
similar to phgen.c) then series of conditional checks to treat the impacts
in order (NXB or source) before saving them in impact file. NXB photons
are NOT imaged through the mirror. Rest of xifupipeline unchanged

--> NXB photon generation in phabkg.c

--> Currently setting NXB photons to a negative PH_ID and SRC_ID to flag
them in the event file

--> Small change in the overall process of the phabkg, no more limited
to 2 files but ton MAX_PHABKG, defined in phabkg.h

--> Modifiying make files to account for new routines (impact.c,
linkedimplist.c...)

--> Adding progressbar.h path to AM_CPPFLAGS in ALL tools to solve
dependencies problems