[ero_calevents] changes RAWX orientatation (needs XML file update!)
The RAWX, RAWY (or rather X,y) coordinate of the eROSITA CCDs is
different to the Sixte standard; this has been corrected with this
commit by flipping the RAWX coordinate in the ero_calevents step.
Therefore the ero RAWX, RAWY are by definition different to the
standard RAWX, RAWY.

As in this case also the rotation angle of the CCDs has to be defined
in the negative order, an update of the eROSITA instrument/XML files
to v1.6.0 is necessary to implement the correct change as defined
and confirmed by Georg Lamer.