erosim: fix bug in initialization of summed ARF
erosim sums the ARFs of all seven telescopes for the photon
generation. The empty array was allocated but not initialized with
value 0. It caused some ARF channels to have very large, negative
values, which then caused the cumulative spectrum to be negative,
causing an error in
erosim: datahandling.c:1466: getSimputPhotonTime: Assertion avgrate>0.' failed.